Move new specs back to utils.tests.js
