+++ lib/converters/utils.test.js
@@ -17,6 +17,412 @@ function expectEntity(entity, entity_id, id, name, type) {
 }
 
 describe('converters/utils', function () {
+    describe("processCommon", function () {
+        beforeEach(function () {
+            entityData.lang = "en";
+        });
+
+        describe("entity ID", function () {
+            it("from ioBroker object ID", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    _id: "ioBroker object ID",
+                };
+
+                const entity = utils.processCommon(null, null, null, obj, entityType);
+
+                expect(entity.entity_id).to.equal(
+                    `${entityType}.${obj._id.replace(/\s/g, "_")}`
+                );
+            });
+
+            it("from common.name", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: "a common name",
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(null, null, null, obj, entityType);
+
+                expect(entity.entity_id).to.equal(
+                    `${entityType}.${obj.common.name.replace(/\s/g, "_")}`
+                );
+            });
+
+            it("from English common.name", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: {
+                            en: "a common name",
+                            de: "ignored",
+                        },
+                    },
+                    _id: "ignored",
+                };
+
+                entityData.lang = "de";
+
+                const entity = utils.processCommon(null, null, null, obj, entityType);
+
+                expect(entity.entity_id).to.equal(
+                    `${entityType}.${obj.common.name.en.replace(/\s/g, "_")}`
+                );
+            });
+
+            it("from instance-specific name", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: "ignored",
+                        custom: {
+                            [entityData.adapter.namespace]: {
+                                name: "instance-specific name",
+                            },
+                        },
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.entity_id).to.equal(
+                    `${entityType}.${obj.common.custom[
+                        entityData.adapter.namespace
+                    ].name.replace(/\W/g, "_")}`
+                );
+            });
+
+            xit("from predefined entity ID argument", function () {
+                const entityType = "does-not-matter";
+                const predefinedEntityId = "predefined id";
+
+                const obj = {
+                    common: {
+                        // TODO
+                        name: "used as the 'idPart' but why when predefinedEntityId is given?",
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    null,
+                    null,
+                    obj,
+                    entityType,
+                    predefinedEntityId
+                );
+
+                expect(entity.entity_id).to.equal(`${entityType}.${obj.common.name}`);
+            });
+
+            it("from predefined entity ID argument with dots", function () {
+                const entityType = "does-not-matter";
+                const predefinedEntityId = "entity-type.predefined id";
+
+                const obj = {
+                    common: {
+                        name: "ignored",
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    null,
+                    null,
+                    obj,
+                    entityType,
+                    predefinedEntityId
+                );
+
+                expect(entity.entity_id).to.equal(
+                    predefinedEntityId.replace(/\s/g, "_")
+                );
+            });
+
+            it("from predefined entity ID argument with illegal characters", function () {
+                const entityType = "does-not-matter";
+                const predefinedEntityId =
+                    "entity-type.predefined id # is an illegal char";
+                const expectedPredefinedEntityId =
+                    "entity-type.predefined_id_is_an_illegal_char";
+
+                const obj = {
+                    common: {
+                        name: "ignored",
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    null,
+                    null,
+                    obj,
+                    entityType,
+                    predefinedEntityId
+                );
+
+                expect(entity.entity_id).to.equal(expectedPredefinedEntityId);
+            });
+        });
+
+        describe("friendly (display) name", function () {
+            it("from ioBroker object ID", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    _id: "ioBroker object ID",
+                };
+
+                const entity = utils.processCommon(null, null, null, obj, entityType);
+
+                expect(entity.attributes.friendly_name).to.equal(obj._id);
+            });
+
+            it("from room and function", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    "room",
+                    "function",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal("room function");
+            });
+
+            it("from common.name", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: "a common name",
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(obj.common.name);
+            });
+
+            it("from common.name in English fallback language", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: {
+                            en: "English",
+                            de: "German",
+                        },
+                    },
+                    _id: "ignored",
+                };
+
+                entityData.lang = undefined;
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(obj.common.name.en);
+            });
+
+            it("from common.name in instance-specific language", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: {
+                            en: "English",
+                            de: "German",
+                        },
+                    },
+                    _id: "ignored",
+                };
+
+                entityData.lang = "de";
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(obj.common.name.de);
+            });
+
+            it("from common.smartName", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: "ignored",
+                        smartName: "a smart name",
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(
+                    obj.common.smartName
+                );
+            });
+
+            it("from common.smartName in fallback language", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: "ignored",
+                        smartName: {
+                            en: "English",
+                            de: "German",
+                        },
+                    },
+                    _id: "ignored",
+                };
+
+                entityData.lang = undefined;
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(
+                    obj.common.smartName.en
+                );
+            });
+
+            it("from common.smartName in instance-specific language", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                    name: "ignored",
+                    smartName: {
+                        en: "English",
+                        de: "German",
+                    },
+                    },
+                    _id: "ignored",
+                };
+
+                entityData.lang = "de";
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(
+                    obj.common.smartName.de
+                );
+            });
+
+            it("ignores name of manually created entity", function () {
+                const entityType = "does-not-matter";
+
+                const obj = {
+                    common: {
+                        name: "ignored",
+                        smartName: "should be used",
+                        custom: {
+                            [entityData.adapter.namespace]: {
+                                name: "should be ignored",
+                            },
+                        },
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    null,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(
+                    obj.common.smartName
+                );
+            });
+
+            it("from name argument", function () {
+                const entityType = "does-not-matter";
+                const predefinedName = "predefined name";
+
+                const obj = {
+                    common: {
+                        name: "ignored",
+                        smartName: "ignored",
+                    },
+                    _id: "ignored",
+                };
+
+                const entity = utils.processCommon(
+                    predefinedName,
+                    "ignored",
+                    "ignored",
+                    obj,
+                    entityType
+                );
+
+                expect(entity.attributes.friendly_name).to.equal(predefinedName);
+            });
+        });
+    });
+
+
     it('processCommon -> id from common.name', function () {
         const id = 'test.0.socket';
         const entity_id = 'switch.Some_Socket';
